PEAR-1838 - cleanup facet code a bit and make it more generic
