+++ packages/portal-proto/src/features/projectsCenter/hooks.ts
@@ -1,26 +1,12 @@
 import {
   ClearFacetFunction,
-  EnumFacetResponse,
   UpdateFacetFilterFunction,
 } from "@/features/facets/types";
 import {
   useAppDispatch,
   useAppSelector,
 } from "@/features/projectsCenter/appApi";
-import {
-  EnumOperandValue,
-  FacetBuckets,
-  fetchFacetByNameGQL,
-  FilterSet,
-  GQLDocType,
-  GQLIndexType,
-  OperandValue,
-  Operation,
-  useCoreDispatch,
-  useCoreSelector,
-  usePrevious,
-  selectProjectsFacetByField,
-} from "@gff/core";
+import { FilterSet, OperandValue, Operation } from "@gff/core";
 import {
   removeProjectFilter,
   selectProjectFiltersByName,
@@ -28,79 +14,8 @@ import {
   updateProjectFilter,
   clearProjectFilters,
 } from "@/features/projectsCenter/projectCenterFiltersSlice";
-import { useCallback, useEffect } from "react";
-import isEqual from "lodash/isEqual";
+import { useCallback } from "react";
 import { extractValue } from "@/features/facets/hooks";
-
-export const useLocalFilters = (
-  field: string,
-  docType: GQLDocType,
-  indexType: GQLIndexType,
-  selectFieldEnumValues: (field: string) => OperandValue,
-  selectLocalFilters: () => FilterSet,
-): EnumFacetResponse => {
-  const coreDispatch = useCoreDispatch();
-
-  const facet: FacetBuckets = useCoreSelector((state) =>
-    selectProjectsFacetByField(state, field),
-  ); // Facet data is always cached in the coreState
-
-  const enumValues = selectFieldEnumValues(field);
-  const allFilters = selectLocalFilters();
-  const prevAllFilters = usePrevious(allFilters);
-  const prevEnumValues = usePrevious(enumValues);
-  console.log({ enumValues, allFilters }, "project");
-
-  useEffect(() => {
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const selectProjectFilters = (_ignore) => allFilters;
-    if (
-      !facet ||
-      !isEqual(prevAllFilters, allFilters) ||
-      !isEqual(prevEnumValues, enumValues)
-    ) {
-      coreDispatch(
-        // pass selectCohortAndRepositoryFilters to fetchFacetByNameGQL to
-        // include both the cohort and local Repository filters.
-        // This is an example of cohort centric + local filters
-        fetchFacetByNameGQL({
-          field: field,
-          docType: docType,
-          index: indexType,
-          // eslint-disable-next-line @typescript-eslint/no-unused-vars
-          caseFilterSelector: (_ignored) => {
-            return {
-              mode: "and",
-              root: {},
-            };
-          },
-          localFilters: allFilters,
-        }),
-      );
-    }
-  }, [
-    coreDispatch,
-    facet,
-    field,
-    allFilters,
-    docType,
-    indexType,
-    prevAllFilters,
-    prevEnumValues,
-    enumValues,
-  ]);
-
-  console.log({ facet });
-  return {
-    data: facet?.buckets,
-    enumFilters: (enumValues as EnumOperandValue)?.map((x) => x.toString()),
-    error: facet?.error,
-    isUninitialized: facet === undefined,
-    isFetching: facet?.status === "pending",
-    isSuccess: facet?.status === "fulfilled",
-    isError: facet?.status === "rejected",
-  };
-};
 
 export const useUpdateProjectsFacetFilter = (): UpdateFacetFilterFunction => {
   const dispatch = useAppDispatch();
